+++ FireAxe.Core/AddonProblem.cs
@@ -4,10 +4,12 @@ namespace FireAxe;
 
 public abstract class AddonProblem : Problem
 {
+    private readonly AddonProblemSource _problemSource;
+
     public AddonProblem(AddonProblemSource problemSource) : base(problemSource)
     {
-        Addon = problemSource.Addon; 
+        _problemSource = problemSource;
     }
 
-    public AddonNode Addon { get; }
+    public AddonNode Addon => _problemSource.Addon;
 }